backend/requirements.txt
asgiref==3.7.2
BareNecessities==0.2.8
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Django==4.1.12
django-cors-headers==4.3.0
django-phone-field==1.8.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
icalendar==5.0.10
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Mail==2.1.0
MarkupSafe==2.1.3
mongoengine==0.27.0
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7